requirements.txt
torch>=2.0
numpy
pandas
scikit-learn
pyyaml
joblib

# Statistical validation
scipy>=1.9.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Table formatting
tabulate

# Dashboard (optional)
streamlit
plotly
